Added And Applied Check Comparison Data Function
I created a version of the check_data function meant for vectors of countries to check the comparison countries fro missing data to avoid error outputs when low-data comparison countries are selected. I applied it to cross-country comparison and the time trends section of the dashboard.

Hopefully this function will become obsolete when a reactive comparison country selector is finished.